Move tests for utils.subprocess to dedicated file
